[refactor] WebsocketUtil과 util 이름 수정
